Refactor: Improve Explore screen data handling, search, and UI state
This commit introduces significant improvements to the Explore screen, including better data caching for tab switching, implementation of local search functionality, and more robust UI state management for loading and error scenarios.

**Key Changes:**

- **Explore Screen (`ExploreFragment.kt`):**
    - **Tab-Specific Data Caching:**
        - `recommendedList`, `popularList`, and `allList` now cache their respective data in static variables within a companion object.
        - `isDataLoaded` flag tracks if initial data for tabs has been fetched.
        - When switching tabs, if data is already cached for that tab, it's displayed immediately from the cache instead of making a new API call. The loader is only shown if the tab's data hasn't been fetched yet.
    - **Local Search Functionality:**
        - A `TextWatcher` is added to the search input (`bind.search`).
        - As the user types, `searchQuery` is updated, and `filterCurrentTabData()` is called.
        - `filterCurrentTabData()` filters the currently displayed list (Recommended, Popular, or All) based on the `searchQuery`, matching against category names (case-insensitive).
        - UI updates:
            - Recycler view visibility is tied to whether the filtered list is empty or not.
            - "No Data" view (`bind.noData`) is shown if the search yields no results.
            - "No Internet" view is shown if the list is empty and the search query is also empty (implies an initial load issue).
    - **Improved Data Loading and UI State:**
        - `handleCategoryResponse()` centralizes logic for processing API responses for categories.
        - It updates the appropriate cached list (`recommendedList`, `popularList`, `allList`) based on `currentSelectedTab`.
        - It correctly manages visibility of loader, swipe-refresh, "No Internet", and "No Data" views based on API call success/failure and data presence.
        - Initial tab selection (Recommended) now correctly triggers data loading only if `isDataLoaded` is false.
        - Pull-to-refresh re-fetches data for the currently selected tab.
        - No internet retry logic also re-fetches based on the current tab.
    - **API Call Update:**
        - `getCategory` API call can now include a `search` parameter (though not currently used in tab-based fetching, it's available for future use).
- **API and Repository (`ApiInterface.kt`, `DashRepository.kt`, `DashViewModel.kt`):**
    - `ApiInterface.getCategory`, `DashRepository.getCategory`, and `DashViewModel.getCategory` now accept an optional `search` parameter (String).
- **Bids Screen (`BidsFragment.kt`, `fragment_bids.xml`):**
    - `viewModel.fetchBids` now correctly passes the current `page` number as a string.
    - `fragment_bids.xml`:
        - Changed "No Purchases Found" to "No Bids Found" in `NoDataView`.
        - Adjusted RecyclerView constraints and bottom loader layout.
- **Offer Screen (`OfferFragment.kt`):**
    - Improved error handling for `getOffers`:
        - If a network error occurs, the `noInternet` view is shown, and the recycler view is hidden.
        - `swipeRefreshLayout` refreshing state is properly reset on error.
- **Layout Adjustments:**
    - `fragment_explore.xml`:
        - Changed search input ID from `email` to `search`.
        - Adjusted margins and ensured `searchExpandLayout` parallax works as intended.
        - Added `noData` view for displaying "No Results Found" during search.
- **Data Model (`FetchBidResponse.kt`):**
    - `FetchBidResponse.Data.Product.pricing` type changed from `Int?` to `Double?` to support decimal pricing.

These changes enhance the user experience on the Explore screen by making tab navigation faster, adding a useful search feature, and providing clearer feedback during data loading and error states.

+++ app/src/main/java/io/bidswipe/app/network/response/FetchBidResponse.kt
@@ -60,7 +60,7 @@ data class FetchBidResponse(
             @SerializedName("images")
             val images: List<String?>?,
             @SerializedName("pricing")
-            val pricing: Int?,
+            val pricing: Double?,
             @SerializedName("product_show")
             val productShow: String?,
             @SerializedName("purchased_quantity")
